Use LLVM 14 on macOS
As build issues are being encountered with LLVM 15 on macOS, pin to LLVM
14 in this feedstock to workaround those build issues. To ensure
`llvm-openmp` matches pin it as well. Also, as the new `libcxx` will
introduce this problem even with the older compiler toolchain, add
`libcxx` as an explicit dependency in `build` & `host` along with
pinning it.

These build issues still need to be fixed upstream. However, this
workaround should just allow this feedstock to continue to build on
macOS.